requirement.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0